cli/pyproject.toml
[project]
name = "paty"
version = "0.1.0"
description = "PATY — Declarative voice agent deployment on Pipecat"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ruamel.yaml>=0.18",
    "pydantic>=2.0",
    "click>=8.0",
    "rich>=13.0",
    "pipecat-ai[local,silero,openai]",
    "websockets>=12.0",
    "httpx>=0.27",
    "loguru>=0.7",
    "numpy>=1.24",
    "opentelemetry-api>=1.20",
    "opentelemetry-sdk>=1.20",
]

[project.optional-dependencies]
mlx = [
    "pipecat-ai[whisper,silero]",
    "mlx-lm>=0.20",
    "mlx-audio>=0.2",
    "misaki[en]<0.9",
]
cuda = [
    "pipecat-ai[whisper,silero]",
    "llama-cpp-python",
]
cpu = [
    "pipecat-ai[whisper,silero]",
    "llama-cpp-python",
]
tui = ["textual>=0.80"]
otlp = ["opentelemetry-exporter-otlp>=1.20"]
prometheus = ["opentelemetry-exporter-prometheus>=0.50b0"]
eject = ["jinja2>=3.0"]
dev = [
    "pytest>=7.4",
    "pytest-asyncio>=0.21",
    "ruff>=0.1.0",
]

[project.scripts]
paty = "paty.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["paty"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "B", "A", "C4", "UP", "SIM", "RUF"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["paty"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
